Fix failing tests from tutorial samples
