+++ codeSnippets/gradle.properties
@@ -10,7 +10,7 @@ ktor_version = 2.3.12
 kotlinx_coroutines_version = 2.0.0
 kotlinx_serialization_version = 2.0.0
 kotlin_css_version = 1.0.0-pre.721
-exposed_version = 0.41.1
+exposed_version = 0.53.0
 h2_version = 2.2.224
 postgresql_version = 42.5.1
 hikaricp_version = 5.1.0
@@ -32,4 +32,4 @@ dropwizard_version = 4.2.25
 bootstrap_version = 5.2.3
 swagger_codegen_version = 1.0.47
 org.gradle.jvmargs = -Xmx2g -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
-
+postgres_version = 42.5.1
