+++ codeSnippets/snippets/tutorial-server-db-integration/src/main/kotlin/com/example/Application.kt
@@ -1,5 +1,6 @@
 package com.example
 
+import com.example.model.FakeTaskRepository
 import com.example.plugins.*
 import io.ktor.server.application.*
 import com.example.model.PostgresTaskRepository
@@ -13,5 +14,12 @@ fun Application.module() {
 
     configureSerialization(repository)
     configureDatabases()
+    configureRouting()
+}
+
+fun Application.testModule() {
+    val repository = FakeTaskRepository()
+
+    configureSerialization(repository)
     configureRouting()
 }
