+++ codeSnippets/snippets/tutorial-server-db-integration/src/main/kotlin/com/example/model/Task.kt
@@ -1,3 +1,5 @@
+package com.example.model
+
 import kotlinx.serialization.Serializable
 
 enum class Priority {
