+++ codeSnippets/snippets/tutorial-server-db-integration/src/main/kotlin/com/example/model/TaskRepository.kt
@@ -1,8 +1,5 @@
 package com.example.model
 
-import Priority
-import Task
-
 interface TaskRepository {
     suspend fun allTasks(): List<Task>
     suspend fun tasksByPriority(priority: Priority): List<Task>
