+++ codeSnippets/snippets/tutorial-server-db-integration/src/main/kotlin/com/example/plugins/Databases.kt
@@ -16,7 +16,6 @@ fun Application.configureDatabases() {
         password = "password"
     )
 }
-
 /**
  * Makes a connection to a Postgres database.
  *
