+++ codeSnippets/snippets/tutorial-server-db-integration/src/main/kotlin/com/example/plugins/Serialization.kt
@@ -1,6 +1,7 @@
 package com.example.plugins
 
-import Task
+import com.example.model.Priority
+import com.example.model.Task
 import com.example.model.TaskRepository
 import io.ktor.http.*
 import io.ktor.serialization.*
