+++ codeSnippets/snippets/tutorial-server-db-integration/src/test/kotlin/com/example/ApplicationTest.kt
@@ -1,24 +1,18 @@
 package com.example
 
-import Task
-import com.example.plugins.*
+import com.example.model.Priority
+import com.example.model.Task
 import io.ktor.client.call.*
 import io.ktor.client.plugins.contentnegotiation.*
 import io.ktor.client.request.*
 import io.ktor.http.*
-import io.ktor.serialization.kotlinx.json.json
+import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.testing.*
 import kotlin.test.*
 
 class ApplicationTest {
     @Test
     fun tasksCanBeFoundByPriority() = testApplication {
-        application {
-            val repository = FakeTaskRepository()
-            configureSerialization(repository)
-            configureRouting()
-        }
-
         val client = createClient {
             install(ContentNegotiation) {
                 json()
@@ -43,20 +37,12 @@ class ApplicationTest {
 
     @Test
     fun unusedPriorityProduces404() = testApplication {
-        application {
-            module()
-        }
-
         val response = client.get("/tasks/byPriority/Vital")
         assertEquals(HttpStatusCode.NotFound, response.status)
     }
 
     @Test
     fun newTasksCanBeAdded() = testApplication {
-        application {
-            module()
-        }
-
         val client = createClient {
             install(ContentNegotiation) {
                 json()
