--- codeSnippets/snippets/tutorial-server-db-integration/src/test/kotlin/com/example/FakeTaskRepository.kt
@@ -1,35 +1,0 @@
-package com.example
-
-import Priority
-import Task
-import com.example.model.TaskRepository
-
-class FakeTaskRepository : TaskRepository {
-    private val tasks = mutableListOf(
-        Task("cleaning", "Clean the house", Priority.Low),
-        Task("gardening", "Mow the lawn", Priority.Medium),
-        Task("shopping", "Buy the groceries", Priority.High),
-        Task("painting", "Paint the fence", Priority.Medium)
-    )
-
-    override suspend fun allTasks(): List<Task> = tasks
-
-    override suspend fun tasksByPriority(priority: Priority) = tasks.filter {
-        it.priority == priority
-    }
-
-    override suspend fun taskByName(name: String) = tasks.find {
-        it.name.equals(name, ignoreCase = true)
-    }
-
-    override suspend fun addTask(task: Task) {
-        if (taskByName(task.name) != null) {
-            throw IllegalStateException("Cannot duplicate task names!")
-        }
-        tasks.add(task)
-    }
-
-    override suspend fun removeTask(name: String): Boolean {
-        return tasks.removeIf { it.name == name }
-    }
-}
