+++ codeSnippets/snippets/tutorial-server-db-integration/src/test/resources/application.yaml
@@ -1,0 +1,6 @@
+ktor:
+  application:
+    modules:
+      - com.example.ApplicationKt.testModule
+  deployment:
+    port: 8080
