--- codeSnippets/snippets/tutorial-server-docker-compose/src/test/kotlin/com/example/ApplicationTest.kt
@@ -1,91 +1,0 @@
-package com.example
-
-import Task
-import com.example.plugins.configureRouting
-import com.example.plugins.configureSerialization
-import io.ktor.client.call.*
-import io.ktor.client.plugins.contentnegotiation.*
-import io.ktor.client.request.*
-import io.ktor.http.*
-import io.ktor.serialization.kotlinx.json.*
-import io.ktor.server.testing.*
-import kotlin.test.*
-
-class ApplicationTest {
-    @Test
-    fun tasksCanBeFoundByPriority() = testApplication {
-        application {
-            val repository = FakeTaskRepository()
-            configureSerialization(repository)
-            configureRouting()
-        }
-
-        val client = createClient {
-            install(ContentNegotiation) {
-                json()
-            }
-        }
-
-        val response = client.get("/tasks/byPriority/Medium")
-        val results = response.body<List<Task>>()
-
-        assertEquals(HttpStatusCode.OK, response.status)
-
-        val expectedTaskNames = listOf("gardening", "painting")
-        val actualTaskNames = results.map(Task::name)
-        assertContentEquals(expectedTaskNames, actualTaskNames)
-    }
-
-    @Test
-    fun invalidPriorityProduces400() = testApplication {
-        application {
-            module()
-        }
-
-        val response = client.get("/tasks/byPriority/Invalid")
-        assertEquals(HttpStatusCode.BadRequest, response.status)
-    }
-
-    @Test
-    fun unusedPriorityProduces404() = testApplication {
-        application {
-            module()
-        }
-
-        val response = client.get("/tasks/byPriority/Vital")
-        assertEquals(HttpStatusCode.NotFound, response.status)
-    }
-
-    @Test
-    fun newTasksCanBeAdded() = testApplication {
-        application {
-            module()
-        }
-
-        val client = createClient {
-            install(ContentNegotiation) {
-                json()
-            }
-        }
-
-        val task = Task("swimming", "Go to the beach", Priority.Low)
-        val response1 = client.post("/tasks") {
-            header(
-                HttpHeaders.ContentType,
-                ContentType.Application.Json
-            )
-
-            setBody(task)
-        }
-        assertEquals(HttpStatusCode.NoContent, response1.status)
-
-        val response2 = client.get("/tasks")
-        assertEquals(HttpStatusCode.OK, response2.status)
-
-        val taskNames = response2
-            .body<List<Task>>()
-            .map { it.name }
-
-        assertContains(taskNames, "swimming")
-    }
-}
