+++ codeSnippets/snippets/tutorial-server-web-application/src/main/resources/application.yaml
@@ -1,0 +1,6 @@
+ktor:
+    application:
+        modules:
+            - com.example.ApplicationKt.module
+    deployment:
+        port: 8080
